migrate all command-related interface functions to CommonService
