+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/service/ExecutorServiceTest.java
@@ -57,6 +57,7 @@ import org.apache.dolphinscheduler.dao.mapper.ProcessTaskRelationMapper;
 import org.apache.dolphinscheduler.dao.mapper.ProjectMapper;
 import org.apache.dolphinscheduler.dao.mapper.TaskDefinitionMapper;
 import org.apache.dolphinscheduler.dao.mapper.TaskGroupQueueMapper;
+import org.apache.dolphinscheduler.service.command.CommandService;
 import org.apache.dolphinscheduler.service.process.ProcessService;
 
 import java.util.ArrayList;
@@ -101,6 +102,9 @@ public class ExecutorServiceTest {
 
     @Mock
     private ProcessService processService;
+
+    @Mock
+    private CommandService commandService;
 
     @Mock
     private ProcessDefinitionMapper processDefinitionMapper;
@@ -194,8 +198,8 @@ public class ExecutorServiceTest {
                 .thenReturn(checkProjectAndAuth());
         Mockito.when(processDefinitionMapper.queryByCode(processDefinitionCode)).thenReturn(processDefinition);
         Mockito.when(processService.getTenantForProcess(tenantId, userId)).thenReturn(new Tenant());
-        doReturn(1).when(processService).createCommand(argThat(c -> c.getId() == null));
-        doReturn(0).when(processService).createCommand(argThat(c -> c.getId() != null));
+        doReturn(1).when(commandService).createCommand(argThat(c -> c.getId() == null));
+        doReturn(0).when(commandService).createCommand(argThat(c -> c.getId() != null));
         Mockito.when(monitorService.getServerListFromRegistry(true)).thenReturn(getMasterServersList());
         Mockito.when(processService.findProcessInstanceDetailById(processInstanceId))
                 .thenReturn(Optional.ofNullable(processInstance));
@@ -230,7 +234,7 @@ public class ExecutorServiceTest {
                 Constants.TEST_FLAG_NO,
                 ComplementDependentMode.OFF_MODE);
         Assertions.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
-        verify(processService, times(1)).createCommand(any(Command.class));
+        verify(commandService, times(1)).createCommand(any(Command.class));
 
     }
 
@@ -253,7 +257,7 @@ public class ExecutorServiceTest {
                 Constants.TEST_FLAG_NO,
                 ComplementDependentMode.OFF_MODE);
         Assertions.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
-        verify(processService, times(1)).createCommand(any(Command.class));
+        verify(commandService, times(1)).createCommand(any(Command.class));
 
     }
 
@@ -320,7 +324,7 @@ public class ExecutorServiceTest {
                 Constants.TEST_FLAG_NO,
                 ComplementDependentMode.OFF_MODE);
         Assertions.assertEquals(Status.START_PROCESS_INSTANCE_ERROR, result.get(Constants.STATUS));
-        verify(processService, times(0)).createCommand(any(Command.class));
+        verify(commandService, times(0)).createCommand(any(Command.class));
     }
 
     /**
@@ -342,7 +346,7 @@ public class ExecutorServiceTest {
                 Constants.TEST_FLAG_NO,
                 ComplementDependentMode.OFF_MODE);
         Assertions.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
-        verify(processService, times(1)).createCommand(any(Command.class));
+        verify(commandService, times(1)).createCommand(any(Command.class));
     }
 
     /**
@@ -364,7 +368,7 @@ public class ExecutorServiceTest {
                 Constants.TEST_FLAG_NO,
                 ComplementDependentMode.OFF_MODE);
         Assertions.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
-        verify(processService, times(31)).createCommand(any(Command.class));
+        verify(commandService, times(31)).createCommand(any(Command.class));
 
     }
 
@@ -387,7 +391,7 @@ public class ExecutorServiceTest {
                 Constants.TEST_FLAG_NO,
                 ComplementDependentMode.OFF_MODE);
         Assertions.assertEquals(Status.SUCCESS, result.get(Constants.STATUS));
-        verify(processService, times(15)).createCommand(any(Command.class));
+        verify(commandService, times(15)).createCommand(any(Command.class));
 
     }
 
@@ -411,7 +415,7 @@ public class ExecutorServiceTest {
 
     @Test
     public void testExecuteRepeatRunning() {
-        Mockito.when(processService.verifyIsNeedCreateCommand(any(Command.class))).thenReturn(true);
+        Mockito.when(commandService.verifyIsNeedCreateCommand(any(Command.class))).thenReturn(true);
         Mockito.when(projectService.checkProjectAndAuth(loginUser, project, projectCode, RERUN))
                 .thenReturn(checkProjectAndAuth());
         Map<String, Object> result =
@@ -421,7 +425,7 @@ public class ExecutorServiceTest {
 
     @Test
     public void testOfTestRun() {
-        Mockito.when(processService.verifyIsNeedCreateCommand(any(Command.class))).thenReturn(true);
+        Mockito.when(commandService.verifyIsNeedCreateCommand(any(Command.class))).thenReturn(true);
         Mockito.when(projectService.checkProjectAndAuth(loginUser, project, projectCode, RERUN))
                 .thenReturn(checkProjectAndAuth());
         Map<String, Object> result = executorService.execProcessInstance(loginUser, projectCode,
