+++ dolphinscheduler-master/src/test/java/org/apache/dolphinscheduler/server/master/runner/WorkflowExecuteRunnableTest.java
@@ -35,6 +35,7 @@ import org.apache.dolphinscheduler.server.master.config.MasterConfig;
 import org.apache.dolphinscheduler.server.master.dispatch.executor.NettyExecutorManager;
 import org.apache.dolphinscheduler.service.alert.ProcessAlertManager;
 import org.apache.dolphinscheduler.service.bean.SpringApplicationContext;
+import org.apache.dolphinscheduler.service.command.CommandService;
 import org.apache.dolphinscheduler.service.expand.CuringParamsService;
 import org.apache.dolphinscheduler.service.process.ProcessService;
 
@@ -72,6 +73,8 @@ public class WorkflowExecuteRunnableTest {
 
     private ProcessService processService;
 
+    private CommandService commandService;
+
     private ProcessInstanceDao processInstanceDao;
 
     private MasterConfig config;
@@ -90,6 +93,7 @@ public class WorkflowExecuteRunnableTest {
 
         config = new MasterConfig();
         processService = Mockito.mock(ProcessService.class);
+        commandService = Mockito.mock(CommandService.class);
         processInstanceDao = Mockito.mock(ProcessInstanceDao.class);
         processInstance = Mockito.mock(ProcessInstance.class);
         Map<String, String> cmdParam = new HashMap<>();
@@ -105,7 +109,8 @@ public class WorkflowExecuteRunnableTest {
         NettyExecutorManager nettyExecutorManager = Mockito.mock(NettyExecutorManager.class);
         ProcessAlertManager processAlertManager = Mockito.mock(ProcessAlertManager.class);
         workflowExecuteThread = Mockito.spy(
-                new WorkflowExecuteRunnable(processInstance, processService, processInstanceDao, nettyExecutorManager,
+                new WorkflowExecuteRunnable(processInstance, commandService, processService, processInstanceDao,
+                        nettyExecutorManager,
                         processAlertManager, config, stateWheelExecuteThread, curingGlobalParamsService));
         Field dag = WorkflowExecuteRunnable.class.getDeclaredField("dag");
         dag.setAccessible(true);
