+++ dolphinscheduler-service/src/test/java/org/apache/dolphinscheduler/service/command/CommandServiceImplTest.java
@@ -1,0 +1,198 @@
+package org.apache.dolphinscheduler.service.command;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import org.apache.dolphinscheduler.common.Constants;
+import org.apache.dolphinscheduler.common.enums.*;
+import org.apache.dolphinscheduler.common.utils.DateUtils;
+import org.apache.dolphinscheduler.common.utils.JSONUtils;
+import org.apache.dolphinscheduler.dao.entity.*;
+import org.apache.dolphinscheduler.dao.mapper.CommandMapper;
+import org.apache.dolphinscheduler.dao.mapper.ProcessDefinitionMapper;
+import org.apache.dolphinscheduler.dao.mapper.ScheduleMapper;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.mockito.junit.jupiter.MockitoSettings;
+import org.mockito.quality.Strictness;
+
+import java.util.*;
+
+import static org.apache.dolphinscheduler.common.Constants.CMD_PARAM_RECOVER_PROCESS_ID_STRING;
+import static org.mockito.ArgumentMatchers.anyString;
+
+@ExtendWith(MockitoExtension.class)
+@MockitoSettings(strictness = Strictness.LENIENT)
+class CommandServiceImplTest {
+
+    @InjectMocks
+    private CommandServiceImpl commandService;
+
+    @Mock
+    private CommandMapper commandMapper;
+
+    @Mock
+    private ProcessDefinitionMapper processDefineMapper;
+
+    @Mock
+    private ScheduleMapper scheduleMapper;
+
+    @Test
+    public void testCreateSubCommand() {
+        ProcessInstance parentInstance = new ProcessInstance();
+        parentInstance.setWarningType(WarningType.SUCCESS);
+        parentInstance.setWarningGroupId(0);
+
+        TaskInstance task = new TaskInstance();
+        task.setTaskParams("{\"processDefinitionCode\":10}}");
+        task.setId(10);
+        task.setTaskCode(1L);
+        task.setTaskDefinitionVersion(1);
+
+        ProcessInstance childInstance = null;
+        ProcessInstanceMap instanceMap = new ProcessInstanceMap();
+        instanceMap.setParentProcessInstanceId(1);
+        instanceMap.setParentTaskInstanceId(10);
+        Command command;
+
+        // father history: start; child null == command type: start
+        parentInstance.setHistoryCmd("START_PROCESS");
+        parentInstance.setCommandType(CommandType.START_PROCESS);
+        ProcessDefinition processDefinition = new ProcessDefinition();
+        processDefinition.setCode(10L);
+        Mockito.when(processDefineMapper.queryByDefineId(100)).thenReturn(processDefinition);
+        Mockito.when(processDefineMapper.queryByCode(10L)).thenReturn(processDefinition);
+        command = commandService.createSubProcessCommand(parentInstance, childInstance, instanceMap, task);
+        Assertions.assertEquals(CommandType.START_PROCESS, command.getCommandType());
+
+        // father history: start,start failure; child null == command type: start
+        parentInstance.setCommandType(CommandType.START_FAILURE_TASK_PROCESS);
+        parentInstance.setHistoryCmd("START_PROCESS,START_FAILURE_TASK_PROCESS");
+        command = commandService.createSubProcessCommand(parentInstance, childInstance, instanceMap, task);
+        Assertions.assertEquals(CommandType.START_PROCESS, command.getCommandType());
+
+        // father history: scheduler,start failure; child null == command type: scheduler
+        parentInstance.setCommandType(CommandType.START_FAILURE_TASK_PROCESS);
+        parentInstance.setHistoryCmd("SCHEDULER,START_FAILURE_TASK_PROCESS");
+        command = commandService.createSubProcessCommand(parentInstance, childInstance, instanceMap, task);
+        Assertions.assertEquals(CommandType.SCHEDULER, command.getCommandType());
+
+        // father history: complement,start failure; child null == command type: complement
+
+        String startString = "2020-01-01 00:00:00";
+        String endString = "2020-01-10 00:00:00";
+        parentInstance.setCommandType(CommandType.START_FAILURE_TASK_PROCESS);
+        parentInstance.setHistoryCmd("COMPLEMENT_DATA,START_FAILURE_TASK_PROCESS");
+        Map<String, String> complementMap = new HashMap<>();
+        complementMap.put(Constants.CMDPARAM_COMPLEMENT_DATA_START_DATE, startString);
+        complementMap.put(Constants.CMDPARAM_COMPLEMENT_DATA_END_DATE, endString);
+        parentInstance.setCommandParam(JSONUtils.toJsonString(complementMap));
+        command = commandService.createSubProcessCommand(parentInstance, childInstance, instanceMap, task);
+        Assertions.assertEquals(CommandType.COMPLEMENT_DATA, command.getCommandType());
+
+        JsonNode complementDate = JSONUtils.parseObject(command.getCommandParam());
+        Date start = DateUtils.stringToDate(complementDate.get(Constants.CMDPARAM_COMPLEMENT_DATA_START_DATE).asText());
+        Date end = DateUtils.stringToDate(complementDate.get(Constants.CMDPARAM_COMPLEMENT_DATA_END_DATE).asText());
+        Assertions.assertEquals(startString, DateUtils.dateToString(start));
+        Assertions.assertEquals(endString, DateUtils.dateToString(end));
+
+        // father history: start,failure,start failure; child not null == command type: start failure
+        childInstance = new ProcessInstance();
+        parentInstance.setCommandType(CommandType.START_FAILURE_TASK_PROCESS);
+        parentInstance.setHistoryCmd("START_PROCESS,START_FAILURE_TASK_PROCESS");
+        command = commandService.createSubProcessCommand(parentInstance, childInstance, instanceMap, task);
+        Assertions.assertEquals(CommandType.START_FAILURE_TASK_PROCESS, command.getCommandType());
+    }
+
+    @Test
+    public void testVerifyIsNeedCreateCommand() {
+
+        List<Command> commands = new ArrayList<>();
+
+        Command command = new Command();
+        command.setCommandType(CommandType.REPEAT_RUNNING);
+        command.setCommandParam("{\"" + CMD_PARAM_RECOVER_PROCESS_ID_STRING + "\":\"111\"}");
+        commands.add(command);
+        Mockito.when(commandMapper.selectList(null)).thenReturn(commands);
+        Assertions.assertFalse(commandService.verifyIsNeedCreateCommand(command));
+
+        Command command1 = new Command();
+        command1.setCommandType(CommandType.REPEAT_RUNNING);
+        command1.setCommandParam("{\"" + CMD_PARAM_RECOVER_PROCESS_ID_STRING + "\":\"222\"}");
+        Assertions.assertTrue(commandService.verifyIsNeedCreateCommand(command1));
+
+        Command command2 = new Command();
+        command2.setCommandType(CommandType.PAUSE);
+        Assertions.assertTrue(commandService.verifyIsNeedCreateCommand(command2));
+    }
+
+    @Test
+    public void testCreateRecoveryWaitingThreadCommand() {
+        int id = 123;
+        Mockito.when(commandMapper.deleteById(id)).thenReturn(1);
+        ProcessInstance subProcessInstance = new ProcessInstance();
+        subProcessInstance.setIsSubProcess(Flag.YES);
+        Command originCommand = new Command();
+        originCommand.setId(id);
+        commandService.createRecoveryWaitingThreadCommand(originCommand, subProcessInstance);
+
+        ProcessInstance processInstance = new ProcessInstance();
+        processInstance.setId(111);
+        commandService.createRecoveryWaitingThreadCommand(null, subProcessInstance);
+
+        Command recoverCommand = new Command();
+        recoverCommand.setCommandType(CommandType.RECOVER_WAITING_THREAD);
+        commandService.createRecoveryWaitingThreadCommand(recoverCommand, subProcessInstance);
+
+        Command repeatRunningCommand = new Command();
+        recoverCommand.setCommandType(CommandType.REPEAT_RUNNING);
+        commandService.createRecoveryWaitingThreadCommand(repeatRunningCommand, subProcessInstance);
+
+        ProcessInstance subProcessInstance2 = new ProcessInstance();
+        subProcessInstance2.setId(111);
+        subProcessInstance2.setIsSubProcess(Flag.NO);
+        commandService.createRecoveryWaitingThreadCommand(repeatRunningCommand, subProcessInstance2);
+    }
+
+    @Test
+    public void giveNullOriginCommand_thenCreateRecoveryWaitingThreadCommand_expectNoDelete() {
+        ProcessInstance subProcessInstance = new ProcessInstance();
+        subProcessInstance.setIsSubProcess(Flag.NO);
+        subProcessInstance.setId(111);
+        ProcessDefinition processDefinition = new ProcessDefinition();
+        processDefinition.setId(111);
+        processDefinition.setCode(10L);
+        subProcessInstance.setProcessDefinition(processDefinition);
+        subProcessInstance.setWarningGroupId(1);
+        commandService.createRecoveryWaitingThreadCommand(null, subProcessInstance);
+        Mockito.verify(commandMapper, Mockito.times(0)).deleteById(anyString());
+    }
+
+    @Test
+    public void testCreateCommand() {
+        Command command = new Command();
+        command.setProcessDefinitionCode(123);
+        command.setCommandParam("{\"ProcessInstanceId\":222}");
+        command.setCommandType(CommandType.START_PROCESS);
+        int mockResult = 1;
+        Mockito.when(commandMapper.insert(command)).thenReturn(mockResult);
+        int exeMethodResult = commandService.createCommand(command);
+        Assertions.assertEquals(mockResult, exeMethodResult);
+        Mockito.verify(commandMapper, Mockito.times(1)).insert(command);
+    }
+
+    @Test
+    public void testFindCommandPageBySlot() {
+        int pageSize = 1;
+        int pageNumber = 0;
+        int masterCount = 0;
+        int thisMasterSlot = 2;
+        List<Command> commandList =
+                commandService.findCommandPageBySlot(pageSize, pageNumber, masterCount, thisMasterSlot);
+        Assertions.assertEquals(0, commandList.size());
+    }
+
+}
